Rename Octopus::Association -> Octopus::AssociationShardTracking
Generally speaking, most modules named Octopus::Foo are then included
directly into ActiveRecord::Foo. Octopus::Association was different from
all of our other modules in that regard, so I gave it a name that does
not map to any of the ActiveRecord::* namespaces.

This was also needed so that I could add a new module Octopus::Association
(for inclusion into the ActiveRecord::Association::Association class) in a
later commit.