Allow use of Octopus with models with custom connection
For backwards compatibility & configurability, this only takes effect
once a class attribute 'allow_octopus' is explicitly set on the model
which has called octopus_establish_connection.

The use case is a model I have that lives in a separate database,
which I do not normally want to have to think about when using it,
but I would like to be able to call "using(:separate_db_standby)"
with an Octopus-managed shard to access a read slave for the model.