Fix failing calls to `select` on `Octopus::RelationProxy`
`Kernel#select` was called instead of `method_missing`

This fix allows `Octopus::RelationProxy` objects via `Octopus::ScopeProxy
to query for specific fields.

When an `Octopus::RelationProxy` is created via an `Octopus::ScopeProxy`
in the context of a query like `User.where()`, it loses the ability to call
`ActiveRecord::QueryMethods#select` via the proxy.

This is because when the `Ocotpus::RelationProxy` tries to call `select`
it incorrectly finds `Kernel#select`.

The solution was to simply add a `select` method on `Octopus::RelationProxy`
to manually dispatch to `method_missing`

See:
 * http://guides.rubyonrails.org/active_record_querying.html#selecting-specific-fields
 * http://api.rubyonrails.org/classes/ActiveRecord/QueryMethods.html
 * http://www.ruby-doc.org/core-2.1.2/Kernel.html#method-i-select