Fix association shard tracking combined with enum method chain
Currently, when Enumerable methods that takes block are called for
relations taken from a shard, ActiveRecord::Relation object is
populated out of the shard, so the records from the master are used.

e.g.:

    User.using(:brazil).first.projects.where(id: ids).map(&:name)

will execute the query for `projects` on the `:master` shard,
instead of `:brazil`.

This patch lists up the Enumerable methods that are not redefined
by ActiveRecord classes, and redirect them to objects taken from
the correct shards.